* origin/master: (104 commits)
  Allow plugins to accept more arguments (#333)
  Bump pytest from 5.4.2 to 5.4.3 (#332)
  Bump pytest-randomly from 3.3.1 to 3.4.0 (#331)
  Bump flake8 from 3.8.1 to 3.8.2 (#330)
  Bump pytest-cov from 2.8.1 to 2.9.0 (#329)
  Bump setupmeta from 2.7.7 to 2.7.8 (#326)
  Bump typing-inspect from 0.5.0 to 0.6.0 (#311)
  Bump lxml from 4.5.0 to 4.5.1 (#323)
  Fix test failing because of loguru upgrade (#325)
  Specify posargs for tox
  Bump tox from 3.15.0 to 3.15.1 (#324)
  Bump loguru from 0.4.1 to 0.5.0 (#322)
  Bump more-itertools from 8.2.0 to 8.3.0 (#321)
  Use 3.8.3
  Bump flake8 from 3.7.9 to 3.8.1 (#318)
  Bump multidict from 4.7.5 to 4.7.6 (#319)
  Bump setupmeta from 2.7.4 to 2.7.7 (#320)
  Bump flake8-quotes from 3.0.0 to 3.2.0 (#317)
  Bump pytest from 5.4.1 to 5.4.2 (#315)
  Add flake8-broken-line (#314)
  ...

+++ tests/test_handler_analysis_path_query_arg.py
@@ -61,9 +61,9 @@ operation = next(
 
 def test_signature_mismatch_missing() -> None:
     async def foo(
-            limit: t.Optional[int],
-            page: t.Optional[float],
-            include_extra: t.Optional[bool],
+        limit: t.Optional[int],
+        page: t.Optional[float],
+        include_extra: t.Optional[bool],
     ) -> response.Response:
         ...
 
@@ -98,10 +98,10 @@ def test_signature_all_missing() -> None:
 
 def test_signature_mismatch_bad_type() -> None:
     async def foo(
-            id: bool,
-            limit: t.Optional[int],
-            page: t.Optional[float],
-            include_extra: t.Optional[bool],
+        id: bool,
+        limit: t.Optional[int],
+        page: t.Optional[float],
+        include_extra: t.Optional[bool],
     ) -> response.Response:
         ...
 
@@ -119,10 +119,10 @@ def test_signature_mismatch_bad_type() -> None:
 
 def test_signature_all_bad_type() -> None:
     async def foo(
-            id: float,
-            limit: t.Optional[t.Union[int, float]],
-            page: t.Optional[t.AbstractSet[bool]],
-            include_extra: t.Union[int, str],
+        id: float,
+        limit: t.Optional[t.Union[float, int]],
+        page: t.Optional[t.AbstractSet[bool]],
+        include_extra: t.Union[int, str],
     ) -> response.Response:
         ...
 
@@ -162,10 +162,10 @@ def test_signature_all_bad_type() -> None:
 
 def test_signature_match() -> None:
     async def test_handler(
-            id: str,
-            limit: t.Optional[int],
-            page: t.Optional[float],
-            include_extra: t.Optional[bool],
+        id: str,
+        limit: t.Optional[int],
+        page: t.Optional[float],
+        include_extra: t.Optional[bool],
     ) -> response.Response:
         ...
